refactor: update test code by changes and make code clean
- Player에서 exitNumber PlayerState로 관리하도록 변경하였기에 테스트 코드도 수정
- OpponentTest에서 조건문 하나 더 추가(중복)
- Validation check 메서드명, 변수명 변경
- Opponent에서 새로운 숫자를 생성할 때 초기화
