+++ kotlin-baseball/src/main/kotlin/baseball/util/ValidationChecker.kt
@@ -1,4 +1,4 @@
-package baseball
+package baseball.util
 
 class ValidationChecker {
 
@@ -6,9 +6,9 @@ class ValidationChecker {
 
     fun checkGameNumber(input: String){
         target = input
-        isAllNumber()
-        isLengthValid()
-        isNotDuplicated()
+        numberCheck()
+        lengthCheck()
+        duplicationCheck()
     }
 
     fun checkExitNumber(input: String){
@@ -17,23 +17,23 @@ class ValidationChecker {
         }
     }
 
-    private fun isAllNumber() {
-        val isAllNumber = target.all { it in '1'..'9' }
-        if (!isAllNumber) {
+    private fun numberCheck() {
+        val isInvalid = target.any { it !in '1'..'9' }
+        if (isInvalid) {
             throw IllegalArgumentException("모든 입력값은 1~9 사이의 숫자여야 합니다.")
         }
     }
 
-    private fun isLengthValid() {
-        val isLengthValid = target.length == 3
-        if (!isLengthValid) {
+    private fun lengthCheck() {
+        val isInvalid = target.length != 3
+        if (isInvalid) {
             throw IllegalArgumentException("입력된 숫자는 3자리 수여야 합니다.")
         }
     }
 
-    private fun isNotDuplicated() {
-        val isNotDuplicated = target.length == target.toMutableList().distinct().size
-        if (!isNotDuplicated) {
+    private fun duplicationCheck() {
+        val isInvalid = target.length != target.toMutableList().distinct().size
+        if (isInvalid) {
             throw IllegalArgumentException("중복된 숫자는 입력되면 안됩니다.")
         }
     }
