+++ kotlin-baseball/src/test/kotlin/baseball/util/ValidationCheckerTest.kt
@@ -1,5 +1,6 @@
-package baseball
+package baseball.util
 
+import baseball.main
 import camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest
 import camp.nextstep.edu.missionutils.test.NsTest
 import org.assertj.core.api.Assertions.assertThatNoException
